requirements.txt
Flask
Flask-SQLAlchemy
PyMySQL